TrafficAnalyzer/requirements.txt
appdirs==1.4.4
dill==0.3.7
joblib==1.3.2
kafka-python==2.0.2
lxml==4.9.3
numpy==1.23.5
opencv-python==4.7.0.72
packaging==23.2
pandas==2.1.2
pip==23.3.1
pyelftools==0.30
PyMySQL==1.1.0
pyshark==0.6
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
scikit-learn==1.2.2
scipy==1.11.4
setuptools==68.0.0
six==1.16.0
termcolor==2.4.0
threadpoolctl==3.2.0
tzdata==2023.3
wheel==0.41.2
